moving some functions to utils; datapaht replace with variables from neuroinfer
