+++ neuroinfer/code/BayesianAnalysis.py
@@ -4,8 +4,8 @@ import pandas as pd
 import pickle
 
 from neuroinfer.code.run_bayesian import run_bayesian_analysis_coordinates, run_bayesian_analysis_area
-from neuroinfer import TEMPLATE_FOLDER
-from neuroinfer.code.plot_generator import generate_nifit_mask
+from neuroinfer import TEMPLATE_FOLDER, PKG_FOLDER, DATA_FOLDER, RESULTS_FOLDER
+from neuroinfer.code.utils import generate_nifti_mask
 from nilearn.image import coord_transform
 
 atlas_path=TEMPLATE_FOLDER/'atlases'/'HarvardOxford'/'HarvardOxford-cort-maxprob-thr25-2mm.nii.gz'
@@ -29,23 +29,16 @@ def run_bayesian_analysis_router(cog_list, area, prior_list, x_target, y_target,
         "'b' for difference measure;\n" +
         "'c' for ratio measure.\n" +
         "'d' for z measure.\n")
-    
-    
-
-    script_directory = os.path.dirname(os.path.abspath(__file__))
-    global_path = os.path.dirname(script_directory)
-    results_folder_path = os.path.join(global_path, "results")
-    data_path = os.path.join(global_path, "data")  # Path to the saved_result folder
 
     # Check if results_folder_path exists, if not, create it
     if not os.path.exists(results_folder_path):
         os.makedirs(results_folder_path) 
 
     if not pd.isnull(area) and pd.isnull(x_target) and pd.isnull(y_target) and pd.isnull(z_target):
-        mask,affine=generate_nifit_mask(area,atlas_path)
+        mask,affine=generate_nifti_mask(area, atlas_path)
         affine_inv = np.linalg.inv(affine) #inverse of the affine matrix to convert from MNI coordinates tp voxel coordinates
 
-        dt_papers_nq, xyz_coords = load_or_calculate_variables(data_path, affine_inv)
+        dt_papers_nq, xyz_coords = load_or_calculate_variables(DATA_FOLDER, affine_inv)
 
         results=run_bayesian_analysis_area(cog_list, prior_list, mask,affine_inv, radius, result_df,cm,dt_papers_nq,xyz_coords)
 
@@ -58,10 +51,10 @@ def run_bayesian_analysis_router(cog_list, area, prior_list, x_target, y_target,
 
     elif not np.isnan(x_target) and not np.isnan(y_target) and not np.isnan(z_target) and np.isnan(area):
         # Call run_bayesian_analysis_coordinates if coordinates are not nan and area is nan
-        mask,affine=generate_nifit_mask(1,atlas_path) #area is set to 1 to get the mask at the coordinates
+        mask,affine=generate_nifti_mask(1, atlas_path) #area is set to 1 to get the mask at the coordinates
         affine_inv = np.linalg.inv(affine) #inverse of the affine matrix to convert from MNI coordinates tp voxel coordinates
 
-        dt_papers_nq, xyz_coords = load_or_calculate_variables(data_path, affine_inv)
+        dt_papers_nq, xyz_coords = load_or_calculate_variables(DATA_FOLDER, affine_inv)
 
         x_target, y_target, z_target = coord_transform(x_target, y_target, z_target,affine_inv)
         results=run_bayesian_analysis_coordinates(cog_list, prior_list, x_target, y_target, z_target, radius, result_df, cm,xyz_coords,dt_papers_nq)
